// src/app/models/customer.ts
export class Customer{
    customer_id: number = 0;
    customer_number: string = '';
    customer_name: string = '';
    address: string = '';
    city: string = '';
    state: string = '';
    country: string = '';
    zipcode: string = '';
    email: string = '';
    mobile_number: string = '';
    whatsapp_number: string = '';
}